Big Double/Denormalization Update

+++ plugins/LinuxVST/src/Air2/Air2.h
@@ -106,7 +106,7 @@ private:
 	bool flipB;
 	bool flop;
 	int count;
-	long double postsine;
+	double postsine;
 
 	
     float A;
