+++ plugins/LinuxVST/src/Baxandall/BaxandallProc.cpp
@@ -45,8 +45,8 @@ void Baxandall::processReplacing(float **inputs, float **outputs, VstInt32 sampl
     
     while (--sampleFrames >= 0)
     {
-		long double inputSampleL = *in1;
-		long double inputSampleR = *in2;
+		double inputSampleL = *in1;
+		double inputSampleR = *in2;
 		if (fabs(inputSampleL)<1.18e-37) inputSampleL = fpd * 1.18e-37;
 		if (fabs(inputSampleR)<1.18e-37) inputSampleR = fpd * 1.18e-37;
 		
@@ -59,10 +59,10 @@ void Baxandall::processReplacing(float **inputs, float **outputs, VstInt32 sampl
 		inputSampleR = sin(inputSampleR);
 		//encode Console5: good cleanness
 		
-		long double trebleSampleL;
-		long double bassSampleL;
-		long double trebleSampleR;
-		long double bassSampleR;
+		double trebleSampleL;
+		double bassSampleL;
+		double trebleSampleR;
+		double bassSampleR;
 		
 		if (flip)
 		{
@@ -182,8 +182,8 @@ void Baxandall::processDoubleReplacing(double **inputs, double **outputs, VstInt
     
     while (--sampleFrames >= 0)
     {
-		long double inputSampleL = *in1;
-		long double inputSampleR = *in2;
+		double inputSampleL = *in1;
+		double inputSampleR = *in2;
 		if (fabs(inputSampleL)<1.18e-43) inputSampleL = fpd * 1.18e-43;
 		if (fabs(inputSampleR)<1.18e-43) inputSampleR = fpd * 1.18e-43;
 		
@@ -196,10 +196,10 @@ void Baxandall::processDoubleReplacing(double **inputs, double **outputs, VstInt
 		inputSampleR = sin(inputSampleR);
 		//encode Console5: good cleanness
 		
-		long double trebleSampleL;
-		long double bassSampleL;
-		long double trebleSampleR;
-		long double bassSampleR;
+		double trebleSampleL;
+		double bassSampleL;
+		double trebleSampleR;
+		double bassSampleR;
 		
 		if (flip)
 		{
