+++ plugins/LinuxVST/src/Console7Cascade/Console7Cascade.h
@@ -55,11 +55,11 @@ private:
 	double gainchase;
 	double chasespeed;
 	
-	long double biquadA[15];
-	long double biquadB[15];
-	long double biquadC[15];
-	long double biquadD[15];
-	long double biquadE[15];
+	double biquadA[15];
+	double biquadB[15];
+	double biquadC[15];
+	double biquadD[15];
+	double biquadE[15];
 	
 	uint32_t fpdL;
 	uint32_t fpdR;
