+++ plugins/LinuxVST/src/NotJustAnotherCD/NotJustAnotherCD.h
@@ -53,8 +53,8 @@ private:
     
 	double bynL[13];
 	double bynR[13];
-	long double noiseShapingL;
-	long double noiseShapingR;
+	double noiseShapingL;
+	double noiseShapingR;
 	
 };
